Separated App.js into various modules for better organization
