+++ frontend/Login.css
@@ -1,0 +1,116 @@
+/* Login Styles */
+
+.login_wrap {
+	/* Structure and Settings*/
+	display: inline-block;
+	background: #0c1223;
+	/* Size */
+	height: auto;
+	width: 1000px;
+	/* Margins and Alignment */
+	margin: 50px 50px 50px 50px;
+	text-align: center;
+	/* Border */
+	border-style: solid;
+	border-color: #303436;
+	border-width: 3px;
+	border-radius: 10px;
+}
+
+.login {
+	/* Structure and Settings */
+	display: inline-block;
+	background-color: #1b1b33;
+	/* Size */
+	min-height: 100px;
+	height: auto;
+	width: 900px;
+	/* Margins */
+	margin-top: 0px;
+	margin-left: 0px;
+	/* Border */
+	border-style: solid;
+	border-color: #29c0b0;
+	border-width: 3px;
+	border-radius: 10px;
+}
+
+.login_row:after {
+	/* Structure and Settings */
+	content: "";
+	display: table;
+	clear: both;
+}
+
+.login_info_column {
+	/* Size */
+	width: 50%;
+	/* Margins and Alignments */
+	margin-top: 10px;
+	float: left;
+	text-align: center;
+}
+
+.login_btns_column {
+	/* Size */
+	width: 33.33%;
+	/* Margins and Alignments */
+	margin-top: 10px;
+	float: left;
+	text-align: center;
+}
+
+.login_btn {
+	/* Font */
+	font-family: Arial, Helvetica, sans-serif;
+	font-size: 25px;
+	/* Size */
+	height: 40px;
+	width: 200px;
+	/* Margins and Alignments */
+	margin-bottom: 25px;
+	vertical-align: middle;
+	/* Border */
+	border-color: #000000;
+	border-style: solid;
+	border-radius: 5px;
+}
+
+.login_btn:hover {
+	/* Font */
+	color: #ffffff;
+	/* Settings */
+	background-color: #000000;
+	/* Border */
+	border: #404040;
+	border-style: solid;
+	border-width: 3px;
+}
+
+.login_btn:active {
+	/* Settings */
+	border: aquamarine;
+	border-style: solid;
+	border-width: 4px;
+}
+
+/* General Styles */
+
+.user {
+	/* Fonts */
+	font-family: 'Courier New', Courier, monospace;
+	font-weight: oblique;
+	font-size: 30px;
+	/* Margins */
+	margin-top: 27px;
+}
+
+.empty_space {
+	/* Structure */
+	display: inline-block;
+	/* Size */
+	height: 10px;
+	width: 2030px;
+	/* Margins */
+	margin-left: 0px;
+}
