+++ frontend/Projects.css
@@ -1,0 +1,258 @@
+/* Project Library Styles */
+
+.project_title {
+	/* Font */
+	font-family: 'Courier New';
+	font-weight: bold;
+	font-size: 40px;
+	/* Margins and Alignment */
+	margin-top: 25px;
+	margin-bottom: 25px;
+	margin-left: 25px;
+	text-align: center;
+}
+
+.project_wrap {
+	/* Structure and Settings*/
+	display: inline-block;
+	background: #0c1223;
+	/* Size */
+	height: auto;
+	width: 1800px;
+	/* Margins and Alignment */
+	margin: 50px 50px 50px 50px;
+	text-align: center;
+	/* Border */
+	border-style: solid;
+	border-color: #303436;
+	border-width: 3px;
+	border-radius: 10px;
+}
+
+.project {
+	/* Structure and Settings */
+	display: inline-block;
+	background-color: #1b1b33;
+	/* Size */
+	min-height: 100px;
+	height: auto;
+	width: 1700px;
+	/* Margins */
+	margin-top: 0px;
+	margin-left: 0px;
+	/* Border */
+	border-style: solid;
+	border-color: #29c0b0;
+	border-width: 3px;
+	border-radius: 10px;
+}
+
+.project_column {
+	/* Size */
+	width: 14.28%;
+	/* Margins and Alignments */
+	margin-top: 10px;
+	float: left;
+	text-align: center;
+}
+
+.new_project_column {
+	/* Size */
+	width: 33.33%;
+	/* Margins and Alignments */
+	margin-top: 10px;
+	float: left;
+	text-align: center;
+}
+
+.project_name {
+	/* Fonts */
+	font-family: Arial, Helvetica, sans-serif;
+	font-weight: bold;
+	font-size: 30px;
+	/* Margins */
+	margin-top: 27px;
+}
+
+.hw_description {
+	/* Font */
+	font-family: Arial, Helvetica, sans-serif;
+	font-size: 25px;
+	/* Margins */
+	margin-top: 5px;
+	margin-bottom: 15px;
+}
+
+.hw_select {
+	/* Font */
+	font-size: 20px;
+	/* Size */
+	height: 30px;
+	width: 120px;
+	/* Margins and Alignment */
+	margin-top: -10px;
+	vertical-align: top;
+}
+
+.hw_input {
+	/* Font */
+	font-size: 19px;
+	/* Size */
+	height: 15px;
+	width: 100px;
+	/* Margins and Alignments */
+	padding: 5px 20px;
+	margin-top: 3px;
+	margin-bottom: 10px;
+	vertical-align: top;
+	text-align: center;
+	/* Border */
+	border-color: aquamarine;
+	border-style: solid;
+	border-width: 3px;
+	border-radius: 5px;
+}
+
+.new_project_input {
+	/* Font */
+	font-size: 25px;
+	/* Size */
+	height: 30px;
+	width: 250px;
+	/* Margins and Alignments */
+	padding: 5px 20px;
+	margin-top: 16px;
+	margin-bottom: 10px;
+	vertical-align: middle;
+	text-align: center;
+	/* Border */
+	border-color: aquamarine;
+	border-style: solid;
+	border-width: 3px;
+	border-radius: 5px;
+}
+
+.check_btn {
+	/* Font */
+	font-family: Arial, Helvetica, sans-serif;
+	font-size: 19px;
+	/* Size */
+	height: 30px;
+	width: 150px;
+	/* Margins and Alignments */
+	margin-top: 3px;
+	margin-bottom: 10px;
+	vertical-align: top;
+	/* Border */
+	border-color: #000000;
+	border-style: solid;
+	border-radius: 5px;
+}
+
+.check_btn:hover {
+	/* Font */
+	color: #ffffff;
+	/* Settings */
+	background-color: #000000;
+	/* Border */
+	border: #404040;
+	border-style: solid;
+	border-width: 3px;
+}
+
+.check_btn:active {
+	/* Border */
+	border: aquamarine;
+	border-style: solid;
+	border-width: 4px;
+}
+
+.join_btn {
+	/* Font */
+	font-family: Arial, Helvetica, sans-serif;
+	font-size: 19px;
+	/* Size */
+	height: 30px;
+	width: 150px;
+	/* Margins and Alignments */
+	margin-top: 25px;
+	margin-bottom: 25px;
+	vertical-align: middle;
+	/* Border */
+	border-color: #000000;
+	border-style: solid;
+	border-radius: 5px;
+}
+
+.join_btn:hover {
+	/* Font */
+	color: #ffffff;
+	/* Settings */
+	background-color: #000000;
+	/* Border */
+	border: #404040;
+	border-style: solid;
+	border-width: 3px;
+}
+
+.join_btn:active {
+	/* Settings */
+	border: aquamarine;
+	border-style: solid;
+	border-width: 4px;
+}
+
+.add_project_btn {
+	/* Font */
+	font-family: Arial, Helvetica, sans-serif;
+	font-size: 25px;
+	/* Size */
+	height: 45px;
+	width: 280px;
+	/* Margins and Alignments */
+	margin-top: 18px;
+	margin-bottom: 25px;
+	vertical-align: middle;
+	/* Border */
+	border-color: #000000;
+	border-style: solid;
+	border-radius: 5px;
+}
+
+.add_project_btn:hover {
+	/* Font */
+	color: #ffffff;
+	/* Settings */
+	background-color: #000000;
+	border: #404040;
+	border-style: solid;
+	border-width: 3px;
+}
+
+.add_project_btn:active {
+	/* Settings */
+	border: aquamarine;
+	border-style: solid;
+	border-width: 4px;
+}
+
+/* General Styles */
+
+.user {
+	/* Fonts */
+	font-family: 'Courier New', Courier, monospace;
+	font-weight: oblique;
+	font-size: 30px;
+	/* Margins */
+	margin-top: 27px;
+}
+
+.empty_space {
+	/* Structure */
+	display: inline-block;
+	/* Size */
+	height: 10px;
+	width: 2030px;
+	/* Margins */
+	margin-left: 0px;
+}
